rename zed files and targets for more consistency
