+++ obelisk_ws/src/cpp/hardware/obelisk_zed_cpp/src/zed2_sensors.cpp
@@ -2,7 +2,7 @@
 
 #include "obelisk_ros_utils.h"
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
-#include "zed2_sensors.h"
+#include "zed2_sensors_interface.h"
 
 int main(int argc, char* argv[]) {
     obelisk::utils::SpinObelisk<ObeliskZed2Sensors, rclcpp::executors::MultiThreadedExecutor>(argc, argv,
